# api-doc.yaml
openapi: 3.0.4
info:
  title: Library Reservation API
  description: |-
    API for the library reservation website, allowing users to book study desks and lockers. 
    Endpoints provide functionality to manage reservations, check desk/locker availability, 
    create, cancel, or update reservations. 
    Designed for integration with web to enhance the user experience in the library.

    Some useful links:
    - [The Library Reservation repository]
    - [The source API definition for the Library Reservation]
  termsOfService: http://libraryreservation.com/terms/
  contact:
    email: support@libraryreservation.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /lockers:
    post:
      tags:
        - Lockers
      summary: Add a new locker to the library.
      description: Add a new locker to the library.
      operationId: createNewLocker
      requestBody:
        description: Add a new locker to the library
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Locker'
        required: true
      responses:
        '200':
          description: Locker added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Locker'
                
    get:
      tags:
        - Lockers
      summary: Get all lockers
      description: Retrieve a list of all lockers in the library
      operationId: getAllLockers
      responses:
        '200':
          description: List of lockers returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Locker'
  /lockers/{id}:
    put:
      tags:
        - Lockers
      summary: Update locker availability status by LockerId.
      description: Update the availability status of a locker by LockerId.
      operationId: updateLocker
      parameters:
        - name: id
          in: path
          description: ID of locker to update availability
          required: true
          schema:
            type: string
      requestBody:
        description: Update locker availability status
        content:
          application/json:
            schema:
              type: object
              properties:
                availability:
                  type: boolean
                  example: true
                date:
                  type: string
                  format: date
                  example: "2024-06-01"
        required: true
      responses:
        '200':
          description: Locker with lockerID has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Locker'
        '404':
          description: Locker not found

    get:
      tags:
        - Lockers
      summary: Get locker by LockerId.
      description: Get locker details by LockerId.
      operationId: getLockerByLockerID
      parameters:
        - name: id
          in: path
          description: ID of locker to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Locker details retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Locker'
        '404':
          description: Locker not found

    delete:
      tags:
        - Lockers
      summary: Delete a locker from the library by LockerId.
      description: Delete a locker from the library.
      operationId: deleteLocker
      parameters:
        - name: id
          in: path
          description: ID of locker to delete
          required: true
          schema:
            type: string
      requestBody:
        description: Delete an existing locker from the library
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Locker'
        required: true
      responses:
        '200':
          description: Locker with lockerID has been deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Locker'
        '404':
          description: Locker not found
  /feedback:
    post:
      tags:
        - Feedback
      summary: Submit user feedback.
      description: Submit feedback or support requests from users.
      operationId: createNewFeedback
      requestBody:
        description: User feedback submission
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Feedback'
        required: true
      responses:
        '201':
          description: Create new feedback with email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
    
    get:
      tags:
        - Feedback
      summary: Get all user feedback
      description: Retrieve a list of all user feedback submissions
      operationId: getAllFeedback
      responses:
        '200':
          description: List of feedback returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feedback'
  /feedback/{email}: 
    get:
      tags:
        - Feedback
      summary: Get feedback by user email.
      description: Get feedback submissions by user email.
      operationId: getFeedbackByEmail
      parameters:
        - name: email
          in: path
          description: Email of user to return feedback
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Feedback details retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feedback'
        '404':
          description: Feedback with email not found
components:
  schemas:
...
    Locker:
      properties:
        lockerId:
          type: integer
          example: 1
        availability:
          type: boolean
          example: true
        username:
          type: string
          example: "23521677"
      required:
        - lockerId
        - availability
        - username
    Feedback:
      properties:
        username:
          type: string
          example: "23521677"
        email:
          type: string
          example: "23521677@gm.uit.edu.vn"
        feedbackType:
          type: string
          example: "Suggestion"
        rating:
          type: integer
          example: 5
        message:
          type: string
          example: "Great service!"
      required:
        - username
        - email
        - feedbackType
        - rating
        - message